requirements.txt
flask==2.2.2
flask-bcrypt==1.0.1
pillow==8.4.0
numpy==1.22.3
pycryptodome==3.14.1

